Add new habit static quick action
